Extract SimpleNoteConverter to its own file
