+++ piano-hero/mobile/src/main/java/com/novoda/pianohero/SimpleEndToEndActivity.java
@@ -8,10 +8,6 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
-import java.util.HashMap;
-import java.util.Locale;
-import java.util.Map;
-
 public class SimpleEndToEndActivity extends AppCompatActivity {
 
     private SimpleNotesOutputView outputView;
@@ -52,30 +48,6 @@ public class SimpleEndToEndActivity extends AppCompatActivity {
                 startNewGame();
             }
         });
-    }
-
-    private static class SimpleNoteConverter {
-
-        private static final Map<String, Note> MAP = new HashMap<>();
-
-        static {
-            MAP.put("c", Note.C4);
-            MAP.put("d", Note.D4);
-            MAP.put("e", Note.E4);
-            MAP.put("f", Note.F4);
-            MAP.put("g", Note.G4);
-            MAP.put("a", Note.A4);
-            MAP.put("b", Note.B4);
-        }
-
-        public Note convert(String raw) {
-            Note note = MAP.get(raw.toLowerCase(Locale.US));
-            if (note == null) {
-                throw new IllegalArgumentException("Argument doesn't correspond to simple note: " + raw);
-            } else {
-                return note;
-            }
-        }
     }
 
     private void startNewGame() {
